@@ -3,26 +3,35 @@ apply plugin: 'kotlin-android'
 apply from: 'maven-push.gradle'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 31
 
     defaultConfig {
         minSdkVersion 16
-        targetSdkVersion 28
+        targetSdkVersion 31
         versionCode 1
-        versionName "1.4.2"
+        versionName "1.9.2"
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
 
     buildTypes {
+        debug{
+            buildConfigField "String", "VERSION_NAME", "\"${defaultConfig.versionName}\""
+        }
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            buildConfigField "String", "VERSION_NAME", "\"${defaultConfig.versionName}\""
         }
     }
 
     defaultConfig {
         consumerProguardFiles 'proguard-rules.pro'
+    }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
 }
 
@@ -32,26 +41,39 @@ dependencies {
     implementation 'com.squareup.retrofit2:converter-gson:2.3.0'
     implementation 'com.squareup.retrofit2:retrofit:2.5.0'
     implementation 'com.squareup.tape2:tape:2.0.0-beta1'
-    implementation 'com.squareup.moshi:moshi:1.8.0'
-    implementation 'com.squareup.moshi:moshi-adapters:1.8.0'
-
-    testImplementation 'junit:junit:4.12'
-    testImplementation "org.mockito:mockito-core:2.13.0"
-    androidTestImplementation 'com.squareup.okhttp3:mockwebserver:3.12.0'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    implementation 'com.squareup.moshi:moshi-kotlin:1.9.2'
+    implementation 'com.squareup.moshi:moshi-adapters:1.9.2'
+    implementation "dev.zacsweers.moshisealed:moshi-sealed-reflect:0.1.0"
+    implementation "dev.zacsweers.moshisealed:moshi-sealed-annotations:0.1.0"
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:${versions.kotlin}"
+    implementation "org.jetbrains.kotlin:kotlin-reflect:1.7.21"
 
-    compileOnly "com.google.android.gms:play-services-gcm:16.1.0"
-    compileOnly "com.google.firebase:firebase-core:16.0.9"
-    compileOnly "com.google.firebase:firebase-messaging:18.0.0"
+    compileOnly "com.google.firebase:firebase-messaging:23.4.0"
+    compileOnly "com.google.firebase:firebase-installations:17.2.0"
 
-    implementation 'com.firebase:firebase-jobdispatcher:0.8.5'
+    implementation("androidx.work:work-runtime-ktx:2.7.0")
+    implementation 'androidx.concurrent:concurrent-futures:1.1.0'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.4'
 
-    implementation 'android.arch.lifecycle:extensions:1.1.1'
-    implementation 'com.android.support:support-annotations:28.0.0'
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
+    implementation 'androidx.annotation:annotation:1.7.1'
 
     lintChecks project(':pushnotifications-lint')
+
+    testImplementation 'junit:junit:4.12'
+    testImplementation 'androidx.test:core:1.5.0'
+    testImplementation 'androidx.test.ext:junit:1.1.5'
+    testImplementation 'androidx.test.ext:truth:1.5.0'
+    testImplementation "org.mockito:mockito-core:2.13.0"
+
+    androidTestImplementation 'androidx.test:core:1.5.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'com.squareup.okhttp3:mockwebserver:3.12.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
+
+    testImplementation "net.bytebuddy:byte-buddy:1.8.22"
+    testImplementation "net.bytebuddy:byte-buddy-agent:1.8.22"
 }
 
 repositories {
