@@ -1,13 +1,12 @@
 package com.example.pushnotificationsintegrationtests
 
-import android.support.test.InstrumentationRegistry
-import android.support.test.runner.AndroidJUnit4
-import com.google.firebase.iid.FirebaseInstanceId
-import com.pusher.pushnotifications.PushNotifications
+import androidx.test.platform.app.InstrumentationRegistry
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.google.firebase.installations.FirebaseInstallations
 import com.pusher.pushnotifications.PushNotificationsInstance
 import com.pusher.pushnotifications.SubscriptionsChangedListener
 import com.pusher.pushnotifications.fcm.MessagingService
-import com.pusher.pushnotifications.internal.DeviceStateStore
+import com.pusher.pushnotifications.internal.InstanceDeviceStateStore
 import org.awaitility.core.ConditionTimeoutException
 import org.awaitility.kotlin.await
 import org.awaitility.kotlin.untilNotNull
@@ -32,14 +31,15 @@ const val DEVICE_REGISTRATION_WAIT_SECS: Long = 15 // We need to wait for FCM to
  */
 @RunWith(AndroidJUnit4::class)
 class DeviceRegistrationTest {
+  val context = InstrumentationRegistry.getInstrumentation().targetContext
+  val instanceId = "00000000-1241-08e9-b379-377c32cd1e83"
+  val errolClient = ErrolAPI(instanceId, "http://localhost:8080")
+
   fun getStoredDeviceId(): String? {
-    val deviceStateStore = DeviceStateStore(InstrumentationRegistry.getTargetContext())
+    val deviceStateStore = InstanceDeviceStateStore(InstrumentationRegistry.getInstrumentation().targetContext, instanceId)
     return deviceStateStore.deviceId
   }
 
-  val context = InstrumentationRegistry.getTargetContext()
-  val instanceId = "00000000-1241-08e9-b379-377c32cd1e83"
-  val errolClient = ErrolAPI(instanceId, "http://localhost:8080")
   companion object {
     val errol = FakeErrol(8080)
 
@@ -53,7 +53,7 @@ class DeviceRegistrationTest {
   @Before
   @After
   fun wipeLocalState() {
-    val deviceStateStore = DeviceStateStore(InstrumentationRegistry.getTargetContext())
+    val deviceStateStore = InstanceDeviceStateStore(InstrumentationRegistry.getInstrumentation().targetContext, instanceId)
 
     await.atMost(1, TimeUnit.SECONDS) until {
       assertTrue(deviceStateStore.clear())
@@ -163,14 +163,14 @@ class DeviceRegistrationTest {
     val getDeviceResponse = errolClient.getDevice(storedDeviceId!!)
     assertNotNull(getDeviceResponse)
 
-    val oldToken = errol.storage.devices[storedDeviceId]?.token
+    val oldToken = errol.getInstanceStorage(instanceId).devices[storedDeviceId]?.token
     assertThat(oldToken, `is`(not(equalTo(""))))
 
-    FirebaseInstanceId.getInstance().deleteInstanceId()
+    FirebaseInstallations.getInstance().delete()
 
     await.atMost(DEVICE_REGISTRATION_WAIT_SECS, TimeUnit.SECONDS) until {
       // The server should have the new token now
-      val newToken = errol.storage.devices[storedDeviceId]?.token
+      val newToken = errol.getInstanceStorage(instanceId).devices[storedDeviceId]?.token
       newToken == oldToken && newToken != ""
     }
   }
@@ -283,7 +283,7 @@ class DeviceRegistrationTest {
 
     // We want to run these callbacks from the UI thread as it's more likely to be useful
     // for the customers. Although, we are not going to make any promises at this point
-    val mainThread = InstrumentationRegistry.getTargetContext().mainLooper.thread
+    val mainThread = InstrumentationRegistry.getInstrumentation().targetContext.mainLooper.thread
     assertThat(lastSetOnSubscriptionsChangedListenerCalledThread, `is`(equalTo(mainThread)))
   }
 }
