@@ -2,18 +2,17 @@ package com.pusher.pushnotifications.reporting
 
 import android.content.Context
 import android.content.Intent
-import android.support.v4.content.WakefulBroadcastReceiver
-import com.firebase.jobdispatcher.*
+import androidx.work.*
 import com.google.gson.Gson
 import com.google.gson.JsonSyntaxException
 import com.pusher.pushnotifications.featureflags.FeatureFlag
 import com.pusher.pushnotifications.featureflags.FeatureFlagManager
 import com.pusher.pushnotifications.internal.AppActivityLifecycleCallbacks
-import com.pusher.pushnotifications.internal.DeviceStateStore
+import com.pusher.pushnotifications.internal.InstanceDeviceStateStore
 import com.pusher.pushnotifications.logging.Logger
 import com.pusher.pushnotifications.reporting.api.DeliveryEvent
 
-class FCMMessageReceiver : WakefulBroadcastReceiver() {
+class FCMMessageReceiver : androidx.legacy.content.WakefulBroadcastReceiver() {
   private val gson = Gson()
   private val log = Logger.get(this::class)
 
@@ -24,8 +23,13 @@ class FCMMessageReceiver : WakefulBroadcastReceiver() {
     }
 
     intent?.getStringExtra("pusher")?.let { pusherDataJson ->
-      try {
-        val pusherData = gson.fromJson(pusherDataJson, PusherMetadata::class.java)
+        val pusherData = try {
+          gson.fromJson(pusherDataJson, PusherMetadata::class.java)
+      } catch (_: JsonSyntaxException) {
+          // TODO: Add client-side reporting
+            log.i("Got an invalid pusher message.")
+            return
+      }
         log.i("Got a valid pusher message.")
 
         if (context == null) {
@@ -33,7 +37,7 @@ class FCMMessageReceiver : WakefulBroadcastReceiver() {
           return
         }
 
-        val deviceStateStore = DeviceStateStore(context)
+        val deviceStateStore = InstanceDeviceStateStore(context, pusherData.instanceId)
 
         val deviceId = deviceStateStore.deviceId
         if (deviceId == null) {
@@ -42,29 +46,27 @@ class FCMMessageReceiver : WakefulBroadcastReceiver() {
         }
 
         val reportEvent = DeliveryEvent(
-          publishId = pusherData.publishId,
-          deviceId = deviceId,
-          userId = deviceStateStore.userId,
-          timestampSecs = Math.round(System.currentTimeMillis() / 1000.0),
-          appInBackground = AppActivityLifecycleCallbacks.appInBackground(),
-          hasDisplayableContent = pusherData.hasDisplayableContent,
-          hasData = pusherData.hasData
+                instanceId = pusherData.instanceId,
+                publishId = pusherData.publishId,
+                deviceId = deviceId,
+                userId = deviceStateStore.userId,
+                timestampSecs = Math.round(System.currentTimeMillis() / 1000.0),
+                appInBackground = AppActivityLifecycleCallbacks.appInBackground(),
+                hasDisplayableContent = pusherData.hasDisplayableContent,
+                hasData = pusherData.hasData
         )
 
-        val dispatcher = FirebaseJobDispatcher(GooglePlayDriver(context))
-        val job = dispatcher.newJobBuilder()
-          .setService(ReportingJobService::class.java)
-          .setTag("pusher.delivered.publishId=${pusherData.publishId}")
-          .setConstraints(Constraint.ON_ANY_NETWORK)
-          .setRetryStrategy(RetryStrategy.DEFAULT_EXPONENTIAL)
-          .setLifetime(Lifetime.UNTIL_NEXT_BOOT)
-          .setExtras(ReportingJobService.toBundle(reportEvent))
-          .build()
+        val reportWorker = OneTimeWorkRequest.Builder(ReportingWorker::class.java)
+                .setConstraints(Constraints.Builder()
+                        .setRequiredNetworkType(NetworkType.CONNECTED)
+                        .build())
+                .setInputData(ReportingWorker.toInputData(reportEvent))
+                .build()
 
-        dispatcher.mustSchedule(job)
-      } catch (_: JsonSyntaxException) {
-        // TODO: Add client-side reporting
-      }
+        val workManagerInstance = WorkManager.getInstance(context)
+        workManagerInstance.enqueueUniqueWork("pusher.delivered.publishId=${pusherData.publishId}",
+                ExistingWorkPolicy.KEEP,
+                reportWorker)
     }
   }
 }
