@@ -22,7 +22,7 @@ data class FakeErrolDevice(
 }
 
 data class FakeErrolStorage(
-  val devices: MutableMap<String, FakeErrolDevice>
+  val devices: MutableMap<String, FakeErrolDevice> = mutableMapOf()
 )
 
 data class RegisterDeviceRequest(
@@ -39,7 +39,11 @@ data class SetSubscriptionsRequest(
 )
 
 class FakeErrol(port: Int, private val clusterKey: String = ""): NanoHTTPDRouter(port) {
-  val storage = FakeErrolStorage(mutableMapOf())
+  private val storage = mutableMapOf<String, FakeErrolStorage>()
+
+  fun getInstanceStorage(instanceId: String): FakeErrolStorage {
+    return storage.getOrPut(instanceId, { FakeErrolStorage() })
+  }
 
   init {
     start(NanoHTTPD.SOCKET_READ_TIMEOUT, false)
@@ -49,7 +53,7 @@ class FakeErrol(port: Int, private val clusterKey: String = ""): NanoHTTPDRouter
     post("/instances/{instanceId}/devices/fcm") {
       entity(RegisterDeviceRequest::class) { registerDeviceRequest ->
         val device = FakeErrolDevice.New(registerDeviceRequest.token, mutableSetOf())
-        storage.devices[device.id] = device
+        getInstanceStorage(params["instanceId"]!!).devices[device.id] = device
 
         complete(Response.Status.OK,
             NewDeviceResponse(id = device.id, initialInterestSet = emptySet()))
@@ -57,11 +61,11 @@ class FakeErrol(port: Int, private val clusterKey: String = ""): NanoHTTPDRouter
     }
 
     put("/instances/{instanceId}/devices/fcm/{deviceId}/token") {
-      val device = storage.devices[params["deviceId"]]
+      val device = getInstanceStorage(params["instanceId"]!!).devices[params["deviceId"]]
       if (device != null) {
         entity(RegisterDeviceRequest::class) { registerDeviceRequest ->
           val device = FakeErrolDevice.New(registerDeviceRequest.token, mutableSetOf())
-          storage.devices[params["deviceId"]!!] = device.copy(token = registerDeviceRequest.token)
+          getInstanceStorage(params["instanceId"]!!).devices[params["deviceId"]!!] = device.copy(token = registerDeviceRequest.token)
 
           complete(Response.Status.OK)
         }
@@ -71,7 +75,7 @@ class FakeErrol(port: Int, private val clusterKey: String = ""): NanoHTTPDRouter
     }
 
     get("/instances/{instanceId}/devices/fcm/{deviceId}") {
-      val device = storage.devices[params["deviceId"]]
+      val device = getInstanceStorage(params["instanceId"]!!).devices[params["deviceId"]]
       if (device != null) {
         complete(Response.Status.OK, GetDeviceResponse(id = device.id, userId = device.userId, deviceMetadata = DeviceMetadata("", "")))
       } else {
@@ -80,9 +84,9 @@ class FakeErrol(port: Int, private val clusterKey: String = ""): NanoHTTPDRouter
     }
 
     delete("/instances/{instanceId}/devices/fcm/{deviceId}") {
-      val device = storage.devices[params["deviceId"]]
+      val device = getInstanceStorage(params["instanceId"]!!).devices[params["deviceId"]]
       if (device != null) {
-        storage.devices -= params["deviceId"]!!
+        getInstanceStorage(params["instanceId"]!!).devices -= params["deviceId"]!!
         complete(Response.Status.OK)
       } else {
         complete(Response.Status.OK)
@@ -90,7 +94,7 @@ class FakeErrol(port: Int, private val clusterKey: String = ""): NanoHTTPDRouter
     }
 
     put("/instances/{instanceId}/devices/fcm/{deviceId}/user", fun NanoHTTPDRouter.Request.(): NanoHTTPD.Response {
-      val device = storage.devices[params["deviceId"]]
+      val device = getInstanceStorage(params["instanceId"]!!).devices[params["deviceId"]]
       if (device == null) {
         return complete(Response.Status.NOT_FOUND)
       }
@@ -105,7 +109,7 @@ class FakeErrol(port: Int, private val clusterKey: String = ""): NanoHTTPDRouter
         val claims = Jwts.parser()
             .setSigningKey(Base64.getEncoder().encode(clusterKey.toByteArray()))
             .parseClaimsJws(jwt)
-        storage.devices[params["deviceId"]!!] = device.copy(userId = claims.body.subject)
+        getInstanceStorage(params["instanceId"]!!).devices[params["deviceId"]!!] = device.copy(userId = claims.body.subject)
         complete(Response.Status.OK)
       } catch (e: Exception) {
         complete(Response.Status.BAD_REQUEST)
@@ -113,7 +117,7 @@ class FakeErrol(port: Int, private val clusterKey: String = ""): NanoHTTPDRouter
     })
 
     get("/instances/{instanceId}/devices/fcm/{deviceId}/interests") {
-      val device = storage.devices[params["deviceId"]]
+      val device = getInstanceStorage(params["instanceId"]!!).devices[params["deviceId"]]
       if (device != null) {
         complete(Response.Status.OK, GetInterestsResponse(interests = device.interests))
       } else {
@@ -122,7 +126,7 @@ class FakeErrol(port: Int, private val clusterKey: String = ""): NanoHTTPDRouter
     }
 
     post("/instances/{instanceId}/devices/fcm/{deviceId}/interests/{interest}") {
-      val device = storage.devices[params["deviceId"]]
+      val device = getInstanceStorage(params["instanceId"]!!).devices[params["deviceId"]]
       if (device != null) {
         device.interests.add(params["interest"]!!)
         complete(Response.Status.OK)
@@ -132,7 +136,7 @@ class FakeErrol(port: Int, private val clusterKey: String = ""): NanoHTTPDRouter
     }
 
     delete("/instances/{instanceId}/devices/fcm/{deviceId}/interests/{interest}") {
-      val device = storage.devices[params["deviceId"]]
+      val device = getInstanceStorage(params["instanceId"]!!).devices[params["deviceId"]]
       if (device != null) {
         device.interests.remove(params["interest"]!!)
         complete(Response.Status.OK)
@@ -143,7 +147,7 @@ class FakeErrol(port: Int, private val clusterKey: String = ""): NanoHTTPDRouter
 
     put("/instances/{instanceId}/devices/fcm/{deviceId}/interests") {
       entity(SetSubscriptionsRequest::class) { setSubscriptions ->
-        val device = storage.devices[params["deviceId"]]
+        val device = getInstanceStorage(params["instanceId"]!!).devices[params["deviceId"]]
         if (device != null) {
           device.interests.clear()
           device.interests.addAll(setSubscriptions.interests)
