@@ -1,5 +1,7 @@
 package com.pusher.pushnotifications;
 
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Set;
 
 import android.app.Activity;
@@ -13,7 +15,7 @@ import com.pusher.pushnotifications.fcm.MessagingService;
  */
 public class PushNotifications {
     private static PushNotificationsInstance instance;
-    protected static TokenProvider tokenProvider;
+    protected static Map<String, TokenProvider> tokenProvider = new HashMap<>();
 
     /**
      * Starts the PushNotification client and synchronizes the FCM device token with
@@ -24,7 +26,18 @@ public class PushNotifications {
      * is deemed better for your project.
      */
     public static PushNotificationsInstance start(Context context, String instanceId) {
-        instance = new PushNotificationsInstance(context, instanceId);
+        if (instance == null) {
+            instance = new PushNotificationsInstance(context, instanceId);
+        } else if (!instance.getInstanceId().equals(instanceId)) {
+            String errorMessage =
+                    "PushNotifications.start has been called before with a different instanceId! (before: "
+                            + instance.getInstanceId() + ", now: " + instanceId + ").\n"
+                            + "If you want to use multiple instanceIds use the `PushNotificationsInstance` class directly "
+                            + "e.g. `val pushNotifications1 = PushNotificationsInstance(context, instanceId)`\n"
+                            + "`pushNotifications1.start()`";
+            throw new IllegalStateException(errorMessage);
+        }
+
         instance.start();
         return instance;
     }
