@@ -1,5 +1,8 @@
 package com.pusher.pushnotifications.internal
 
+import com.pusher.pushnotifications.logging.Logger
+import com.squareup.moshi.JsonDataException
+import com.squareup.moshi.JsonEncodingException
 import com.squareup.tape2.ObjectQueue
 import com.squareup.tape2.QueueFile
 import java.io.*
@@ -13,7 +16,25 @@ interface PersistentJobQueue<T: Serializable> {
 }
 
 class TapeJobQueue<T: Serializable>(file: File, converter: ObjectQueue.Converter<T>): PersistentJobQueue<T> {
-  private var queue = ObjectQueue.create( QueueFile.Builder(file).build(), converter )
+  private val log = Logger.get(this::class)
+  private val noExceptionConverter = object : ObjectQueue.Converter<T?> {
+    override fun from(source: ByteArray): T? {
+      return try {
+        converter.from(source)
+      } catch (ex: JsonEncodingException) {
+        log.w("Failed to read object data from tape. Continuing without this data.")
+        null
+      } catch (ex: JsonDataException) {
+        log.w("Failed to read object data from tape. Continuing without this data.")
+        null
+      }
+    }
+
+    override fun toStream(value: T?, sink: OutputStream) {
+      converter.toStream(value!!, sink)
+    }
+  }
+  private val queue = ObjectQueue.create(QueueFile.Builder(file).build(), noExceptionConverter)
 
   @Synchronized override fun push(job: T) {
     queue.add(job)
@@ -32,6 +53,6 @@ class TapeJobQueue<T: Serializable>(file: File, converter: ObjectQueue.Converter
   }
 
   @Synchronized override fun asIterable(): Iterable<T> {
-    return queue.asIterable()
+    return queue.asIterable().filterNotNull()
   }
 }
