@@ -1,32 +1,36 @@
 package com.pusher.pushnotifications.internal
 
 import com.pusher.pushnotifications.api.DeviceMetadata
-import com.squareup.moshi.adapters.PolymorphicJsonAdapterFactory
+import com.squareup.moshi.JsonClass
+import dev.zacsweers.moshisealed.annotations.TypeLabel
 import java.io.Serializable
 
-sealed class ServerSyncJob: Serializable {
+@JsonClass(generateAdapter = false, generator = "sealed:ServerSyncJob")
+sealed class ServerSyncJob: Serializable
 
-    companion object {
-        // the following is required for the Persistent Job Queue to read and write data correctly.
-        // if you add another ServerSyncJob, you must also add it below here too!
-        val polymorphicJsonAdapterFactory: PolymorphicJsonAdapterFactory<ServerSyncJob>
-                = PolymorphicJsonAdapterFactory.of(ServerSyncJob::class.java, "ServerSyncJob")
-                .withSubtype(StartJob::class.java, "StartJob")
-                .withSubtype(RefreshTokenJob::class.java, "RefreshTokenJob")
-                .withSubtype(SubscribeJob::class.java, "SubscribeJob")
-                .withSubtype(UnsubscribeJob::class.java, "UnsubscribeJob")
-                .withSubtype(SetSubscriptionsJob::class.java, "SetSubscriptionsJob")
-                .withSubtype(ApplicationStartJob::class.java, "ApplicationStartJob")
-                .withSubtype(SetUserIdJob::class.java, "SetUserIdJob")
-                .withSubtype(StopJob::class.java, "StopJob")
-    }
-}
-
+// If you add new fields to any of these data classes, any older records on peoples devices may not
+// be read and ignored as they do not match the current expected class.
+// Please see the PersistentJobQueueTests to see the expected behaviour.
+@TypeLabel("StartJob")
 data class StartJob(val fcmToken: String, val knownPreviousClientIds: List<String>): ServerSyncJob()
+
+@TypeLabel("RefreshTokenJob")
 data class RefreshTokenJob(val newToken: String): ServerSyncJob()
+
+@TypeLabel("SubscribeJob")
 data class SubscribeJob(val interest: String): ServerSyncJob()
+
+@TypeLabel("UnsubscribeJob")
 data class UnsubscribeJob(val interest: String): ServerSyncJob()
+
+@TypeLabel("SetSubscriptionsJob")
 data class SetSubscriptionsJob(val interests: Set<String>): ServerSyncJob()
+
+@TypeLabel("ApplicationStartJob")
 data class ApplicationStartJob(val deviceMetadata: DeviceMetadata): ServerSyncJob()
+
+@TypeLabel("SetUserIdJob")
 data class SetUserIdJob(val userId: String): ServerSyncJob()
+
+@TypeLabel("StopJob")
 class StopJob: ServerSyncJob()
