@@ -2,6 +2,8 @@ package com.pusher.pushnotifications.internal
 
 import com.pusher.pushnotifications.api.DeviceMetadata
 import com.squareup.moshi.Moshi
+import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
+import dev.zacsweers.moshisealed.reflect.MoshiSealedJsonAdapterFactory
 import junit.framework.Assert.*
 import org.junit.Before
 import org.junit.Test
@@ -9,7 +11,9 @@ import java.io.File
 
 class PersistentJobQueueTest {
   lateinit var tempFile: File
-  private val moshi = Moshi.Builder().add(ServerSyncJob.polymorphicJsonAdapterFactory).build()
+  private val moshi = Moshi.Builder()
+          .add(MoshiSealedJsonAdapterFactory())
+          .add(KotlinJsonAdapterFactory()).build()
   private val converter = MoshiConverter(moshi.adapter(ServerSyncJob::class.java))
 
   @Before
@@ -72,7 +76,7 @@ class PersistentJobQueueTest {
   @Test
   fun `iterables returns correct Start Job`() {
     val queue: PersistentJobQueue<ServerSyncJob> = TapeJobQueue(tempFile, converter)
-    queue.push(StartJob("fcm_token", arrayListOf("one", "two")))
+    queue.push(StartJob("fcm_token", arrayListOf("phone1", "phone2")))
 
     val returnedList = queue.asIterable().toList()
 
@@ -83,8 +87,8 @@ class PersistentJobQueueTest {
     val startJob = returnedList[0] as StartJob
     assertEquals("fcm_token", startJob.fcmToken)
     assertEquals(2, startJob.knownPreviousClientIds.size)
-    assertEquals("one", startJob.knownPreviousClientIds[0])
-    assertEquals("two", startJob.knownPreviousClientIds[1])
+    assertEquals("phone1", startJob.knownPreviousClientIds[0])
+    assertEquals("phone2", startJob.knownPreviousClientIds[1])
   }
 
   @Test
@@ -105,7 +109,7 @@ class PersistentJobQueueTest {
   @Test
   fun `iterables returns correct Subscribe Job`() {
     val queue: PersistentJobQueue<ServerSyncJob> = TapeJobQueue(tempFile, converter)
-    queue.push(SubscribeJob("interest_subscribe"))
+    queue.push(SubscribeJob("carrots"))
 
     val returnedList = queue.asIterable().toList()
 
@@ -114,13 +118,13 @@ class PersistentJobQueueTest {
 
     assertEquals(SubscribeJob::class.java, returnedList[0].javaClass)
     val subscribeJob = returnedList[0] as SubscribeJob
-    assertEquals("interest_subscribe", subscribeJob.interest)
+    assertEquals("carrots", subscribeJob.interest)
   }
 
   @Test
   fun `iterables returns correct Unsubscribe Job`() {
     val queue: PersistentJobQueue<ServerSyncJob> = TapeJobQueue(tempFile, converter)
-    queue.push(UnsubscribeJob("interest_unsubscribe"))
+    queue.push(UnsubscribeJob("asparagus"))
 
     val returnedList = queue.asIterable().toList()
 
@@ -129,13 +133,13 @@ class PersistentJobQueueTest {
 
     assertEquals(UnsubscribeJob::class.java, returnedList[0].javaClass)
     val unsubscribeJob = returnedList[0] as UnsubscribeJob
-    assertEquals("interest_unsubscribe", unsubscribeJob.interest)
+    assertEquals("asparagus", unsubscribeJob.interest)
   }
 
   @Test
   fun `iterables returns correct Set Subscriptions Job`() {
     val queue: PersistentJobQueue<ServerSyncJob> = TapeJobQueue(tempFile, converter)
-    queue.push(SetSubscriptionsJob(setOf("subscribe_one", "subscribe_two")))
+    queue.push(SetSubscriptionsJob(setOf("apples", "pears")))
 
     val returnedList = queue.asIterable().toList()
 
@@ -144,15 +148,15 @@ class PersistentJobQueueTest {
 
     assertEquals(SetSubscriptionsJob::class.java, returnedList[0].javaClass)
     val unsubscribeJob = returnedList[0] as SetSubscriptionsJob
-    assertEquals("subscribe_one", unsubscribeJob.interests.first())
-    assertEquals("subscribe_two", unsubscribeJob.interests.last())
+    assertEquals("apples", unsubscribeJob.interests.first())
+    assertEquals("pears", unsubscribeJob.interests.last())
   }
 
   @Test
   fun `iterables returns correct Application Start Job`() {
     val queue: PersistentJobQueue<ServerSyncJob> = TapeJobQueue(tempFile, converter)
     queue.push(ApplicationStartJob(
-            DeviceMetadata("sdk_version", "android_version")))
+            DeviceMetadata("18", "5")))
 
     val returnedList = queue.asIterable().toList()
 
@@ -161,14 +165,14 @@ class PersistentJobQueueTest {
 
     assertEquals(ApplicationStartJob::class.java, returnedList[0].javaClass)
     val applicationStartJob = returnedList[0] as ApplicationStartJob
-    assertEquals("sdk_version", applicationStartJob.deviceMetadata.sdkVersion)
-    assertEquals("android_version", applicationStartJob.deviceMetadata.androidVersion)
+    assertEquals("18", applicationStartJob.deviceMetadata.sdkVersion)
+    assertEquals("5", applicationStartJob.deviceMetadata.androidVersion)
   }
 
   @Test
   fun `iterables returns correct Application Set User Id Job`() {
     val queue: PersistentJobQueue<ServerSyncJob> = TapeJobQueue(tempFile, converter)
-    queue.push(SetUserIdJob("user_id"))
+    queue.push(SetUserIdJob("cucas"))
 
     val returnedList = queue.asIterable().toList()
 
@@ -177,20 +181,20 @@ class PersistentJobQueueTest {
 
     assertEquals(SetUserIdJob::class.java, returnedList[0].javaClass)
     val setUserIdJob = returnedList[0] as SetUserIdJob
-    assertEquals("user_id", setUserIdJob.userId)
+    assertEquals("cucas", setUserIdJob.userId)
   }
 
   @Test
   fun `iterables returns correct Server Sync Job Types`() {
     val queue: PersistentJobQueue<ServerSyncJob> = TapeJobQueue(tempFile, converter)
-    queue.push(StartJob("fcm_token", arrayListOf("one", "two")))
+    queue.push(StartJob("fcm_token", arrayListOf("phone1", "phone2")))
     queue.push(RefreshTokenJob("new_token"))
-    queue.push(SubscribeJob("interest_subscribe"))
-    queue.push(UnsubscribeJob("interest_unsubscribe"))
-    queue.push(SetSubscriptionsJob(setOf("subscribe_one", "subscribe_two")))
+    queue.push(SubscribeJob("carrots"))
+    queue.push(UnsubscribeJob("okra"))
+    queue.push(SetSubscriptionsJob(setOf("apples", "pears")))
     queue.push(ApplicationStartJob(
-            DeviceMetadata("sdk_version", "android_version")))
-    queue.push(SetUserIdJob("user_id"))
+            DeviceMetadata("16", "10")))
+    queue.push(SetUserIdJob("danielle"))
     queue.push(StopJob())
 
     val retrievedElements = queue.asIterable().toList()
@@ -206,6 +210,64 @@ class PersistentJobQueueTest {
     assertTrue(retrievedElements[5] is ApplicationStartJob)
     assertTrue(retrievedElements[6] is SetUserIdJob)
     assertTrue(retrievedElements[7] is StopJob)
+  }
+
+
+  @Test
+  fun `corrupted saved data - existing type object has field added`() {
+    val tempFile = File("src/test/resources/com/pusher/pushnotifications/internal/persistentJobQueue-corrupted_existing_object_field_added.queue")
+    val queue: PersistentJobQueue<ServerSyncJob> = TapeJobQueue(tempFile, converter)
+
+      // uncomment the following to write this to the file
+//    queue.push(SubscribeJob("potato", 5)) //added an interest level of type int field
+//    queue.push(UnsubscribeJob("carrot"))
+//    queue.push(UnsubscribeJob("pear"))
+
+      val retrievedElements = queue.asIterable().toList()
+      assertEquals(3, retrievedElements.size)
+
+      assertEquals((retrievedElements.first() as SubscribeJob).interest, "potato")
+      assertEquals((retrievedElements[1] as UnsubscribeJob).interest, "carrot")
+      assertEquals((retrievedElements[2] as UnsubscribeJob).interest, "pear")
+  }
+
+  @Test
+  fun `corrupted saved data - existing type object has field removed`() {
+    val tempFile = File("src/test/resources/com/pusher/pushnotifications/internal/persistentJobQueue-corrupted_existing_object_field_removed.queue")
+    val queue: PersistentJobQueue<ServerSyncJob> = TapeJobQueue(tempFile, converter)
+
+    // uncomment the following to write this to the file
+//    queue.push(StartJob("fcm_token")) //removed the knownPreviousClientIds field
+//    queue.push(UnsubscribeJob("carrot"))
+//    queue.push(UnsubscribeJob("pear"))
+
+    val retrievedElements = queue.asIterable().toList()
+    assertEquals(2, retrievedElements.size)
+
+    // queue element 0 could not be parsed (missing the knownPreviousClientids so the whole record gets ignored
+
+    assertEquals((retrievedElements[0] as UnsubscribeJob).interest, "carrot")
+    assertEquals((retrievedElements[1] as UnsubscribeJob).interest, "pear")
+
+  }
+
+  @Test
+  fun `corrupted saved data - existing type no longer exists`() {
+    val tempFile = File("src/test/resources/com/pusher/pushnotifications/internal/persistentJobQueue-corrupted_existing_type_no_longer_exists.queue")
+    val queue: PersistentJobQueue<ServerSyncJob> = TapeJobQueue(tempFile, converter)
+
+    // uncomment the following to write this to the file
+//    queue.push(DummyJob("dummy_data")) // this data class no longer exists!
+//    queue.push(UnsubscribeJob("carrot"))
+//    queue.push(UnsubscribeJob("pear"))
+
+    val retrievedElements = queue.asIterable().toList()
+    assertEquals(2, retrievedElements.size)
+
+    // queue element 0 could not be parsed (missing the knownPreviousClientids so the whole record gets ignored
+
+    assertEquals((retrievedElements[0] as UnsubscribeJob).interest, "carrot")
+    assertEquals((retrievedElements[1] as UnsubscribeJob).interest, "pear")
   }
 
 }
